refactor: standardize create toast logics
