+++ packages/toast/src/toast.tsx
@@ -8,9 +8,11 @@ import {
 } from "@chakra-ui/alert"
 import { chakra } from "@chakra-ui/system"
 import { CloseButton } from "@chakra-ui/close-button"
-import { isFunction } from "@chakra-ui/utils"
+import { isFunction, MaybeFunction, runIfFn } from "@chakra-ui/utils"
 import type { UseToastOptions } from "./use-toast"
-import type { RenderProps } from "./toast.types"
+import type { RenderProps, ToastId } from "./toast.types"
+import { getToastPlacement } from "./toast.placement"
+import { toastStore } from "./toast.store"
 
 export interface ToastProps
   extends UseToastOptions,
@@ -79,4 +81,67 @@ export function createRenderToast(
     return <ToastComponent {...props} {...options} />
   }
   return renderToast
+}
+
+type UseToastPromiseOption = Omit<UseToastOptions, "status">
+
+export function createToastFn(
+  dir: "ltr" | "rtl",
+  defaultOptions?: UseToastOptions,
+) {
+  const normalizeToastOptions = (options?: UseToastOptions) => ({
+    ...defaultOptions,
+    ...options,
+    position: getToastPlacement(
+      options?.position ?? defaultOptions?.position,
+      dir,
+    ),
+  })
+
+  const toast = (options?: UseToastOptions) => {
+    const normalizedToastOptions = normalizeToastOptions(options)
+    const Message = createRenderToast(normalizedToastOptions)
+    return toastStore.notify(Message, normalizedToastOptions)
+  }
+
+  toast.update = (id: ToastId, options: Omit<UseToastOptions, "id">) => {
+    toastStore.update(id, normalizeToastOptions(options))
+  }
+
+  toast.promise = <Result extends any, Err extends Error = Error>(
+    promise: Promise<Result>,
+    options: {
+      success: MaybeFunction<UseToastPromiseOption, [Result]>
+      error: MaybeFunction<UseToastPromiseOption, [Err]>
+      loading: UseToastPromiseOption
+    },
+  ) => {
+    const id = toast({
+      ...options.loading,
+      status: "loading",
+      duration: null,
+    })
+
+    promise
+      .then((data) =>
+        toast.update(id, {
+          status: "success",
+          duration: 5_000,
+          ...runIfFn(options.success, data),
+        }),
+      )
+      .catch((error) =>
+        toast.update(id, {
+          status: "error",
+          duration: 5_000,
+          ...runIfFn(options.error, error),
+        }),
+      )
+  }
+
+  toast.closeAll = toastStore.closeAll
+  toast.close = toastStore.close
+  toast.isActive = toastStore.isActive
+
+  return toast
 }
